Alpenhorn extensions for CHIME (#1)
These are the CHIME-specific parts of alpenhorn-2.

## import-detect

The import-detect module's entrypoint is `import_detect()` in `detection.py`.  This function re-creates the CHIME alpenhorn detection behaviour as follows:
 * Split the passed-in path into a acq-name and file-name.
 * Validate the `acq-name` (in `parse_acq()`), by ensuring it:
   1. the type name is listed in the `AcqType` table
   2. the insturment name is listed in the `ArchiveInst` table
   3. the acquisition time is valid
 * Validate the `file-name` by running through `FileType`s for the detected `AcqType` (these come from the new `AcqFileTypes` table in `chimedb.data_index), and ensure one of the `FileType`s matches the file-name.
 * If acq-name and file-name validation both succeed, create a new `partial` wrapping the callback function `store_info()` to bind the detection data to the callback.  The callback is called after the new Alpenhorn records are created (`ArchiveAcq, ArchiveFile, ArchiveFileCopy`)
 * In the callback, populate info tables and records the extra fields to `ArchiveAcq` and `ArchiveFile`

## Test suite
The test suite includes some test files which are meant be imported by a standard run of `alpenhorn`. The data files in `testdata` are real files copied from the archive but with most of the datasets which alpenhorn doesn't care about deleted to save on space.