feat(SciNetHPSSIO): Alpenhorn I/O extension for SciNet HPSS
This is, largely, a reimplementation of the special HPSS code
we had in alpenhorn-1, re-constituted as a third-party alpenhorn
I/O extension (because it's too CHIME-specific to be put in alpenhorn
per se).

BASIC OPERATION

We access HPSS though the `hsi` command-line tool.  However, we can't
run `hsi` directly at SciNet.  Rather, it needs to be run from one of
the SLURM "archive" partitions (either `archiveshort`, which is good
for interactive jobs, or `archivelong`, which is what we use).

To do this, what this extension does to interface with HPSS is:
write a bash script with a bunch of `hsi` commands and then submits the
script via SLURM to the archive partition to run it.

TALKBACK FILES

Operational success or failure is indicated by the bash script through
"talkback" files, which are zero-length files which the script creates
in a special directory.  All information in a talkback file is in its
filename, and alpenhorn will periodically look for these files and
update the database based on the talkbacks.

This system is copied directly from alpenhorn-1.  Originally, the
bash scripts could update the database directly, but at some point the
node servicing the archive partition lost access to our database, so
this weird filesystem based interprocess communication was the
workaround we hacked up in short order.

DIFFERENCES FROM ALPENHORN-1

I think, in bulk, this behaves much like the alpenhorn-1 implementation,
but there are a few changes:
* alpenhorn-1 kept no running state: it would periodically decide it was
  time to shove some stuff in HPSS, look around for anything that was
  waiting for a pull, put all those files in a SLURM job, and then set
  the associated ArchiveFileCopyRequest to complete, banking on the fact
  that if the transfer didn't work, someone would come along later and
  re-sync `scinet_staging` to `scinet_hpss` so it could try again.

  Instead of doing that, this implementation maintains lists of
  in-progress files, in two stages: first in a FIFO of files waiting for
  ingest, and then as lists of files being handled in active SLURM jobs.
  The second of these is to stop alpenhorn from queueing the same
  transfer twice (though the system is pretty robust to that, so it
  probably doesn't matter).

  The first part, the wait FIFO is new and allows files to accumulate
  before alpenhorn triggers SLURM job creation.  A job will be sumitted
  if at least one of these is true:
  * there is more than 800 GiB waiting for ingest
  * there are more than 500 files waiting for ingest
  * the eldest file has been waiting for at least a hour
  The first two of these also set the maximum number of files which will
  be processed in a single SLURM job
* Because this implementation tracks files being dealt with in its
  running state, it needs to know when the jobs are done so it can
  discard those files (so they don't get permanently stuck in the
  in-progress queue).  This is a change from alpenhorn-1, which would
  forget everything about a job (even its existence) the moment the job
  was submitted.

INEFFICIENCIES

The main problem is that all the SLURM commands are happening in the
main thread.  It might make sense to move them to a worker, but I'm
going to leave them as-is for now.  Because all the HPSS interactions
happen in a SLURM job, we don't need to be very efficient here, unlike
with fir /nearline, where agressive optimisation is necessary.

See https://github.com/chime-experiment/alpenhorn-chime/issues/11 for
other remaining issues with this implementation.